Bronze Challenge: Displaying a Number Pad
The keyboard for the UITextField that displays a BNRItem’s valueInDollars is a QWERTY keyboard. It would be better if it was a number pad. Change the Keyboard Type of that UITextField to the Number Pad. (Hint: you can do this in the XIB file using the attributes inspector.) ”

Excerpt From: Conway, Joe. “iOS Programming: The Big Nerd Ranch Guide (4th Edition) (Big Nerd Ranch Guides).” iBooks.